improve: The types of Image.byteCount and MemoryCache.Value.size were changed from Int to Long

+++ sample/src/desktopMain/kotlin/com/github/panpf/sketch/sample/ui/components/zoomimage/core/BufferedImageValue.kt
@@ -8,7 +8,7 @@ class BufferedImageValue(
     override val extras: Map<String, Any?> = emptyMap()
 ) : MemoryCache.Value {
 
-    override val size: Int = image.byteCount
+    override val size: Long = image.byteCount
 
     override fun checkValid(): Boolean {
         return image.checkValid()
